+++ R/backend-spark-sql.R
@@ -138,12 +138,8 @@ simulate_spark_sql <- function() simulate_dbi("Spark SQL")
                                    analyze = TRUE,
                                    in_transaction = FALSE) {
 
-  if (temporary) {
-    sql <- sql_values_subquery(con, values, types = types, lvl = 1)
-    db_compute(con, table, sql, overwrite = overwrite)
-  } else {
-    NextMethod()
-  }
+  sql <- sql_values_subquery(con, values, types = types, lvl = 1)
+  db_compute(con, table, sql, overwrite = overwrite, temporary = temporary)
 }
 
 #' @export
@@ -158,14 +154,11 @@ simulate_spark_sql <- function() simulate_dbi("Spark SQL")
                                    analyze = TRUE,
                                    in_transaction = FALSE) {
 
-  if (!temporary) {
-    cli::cli_abort("Spark SQL only support temporary tables")
-  }
-
   sql <- glue_sql2(
     con,
     "CREATE ", if (overwrite) "OR REPLACE ",
-    "TEMPORARY VIEW {.tbl {table}} AS \n",
+    if (temporary) "TEMPORARY VIEW" else "TABLE",
+    " {.tbl {table}} AS \n",
     "{.from {sql}}"
   )
   DBI::dbExecute(con, sql)