+++ NEWS.md
@@ -3,6 +3,8 @@
 * `across(everything())` doesn't select grouping columns created via `.by` in
   `summarise()` (@mgirlich, #1493).
 *  New translations of clock function `date_count_between()` for SQL server, Redshift, Snowflake, Postgres, and Spark (@edward-burn, #1495).
+* Spark SQL backend now supports persisting tables with
+  `compute(x, name = I("x.y.z"), temporary = FALSE)` (@zacdav-db, #1502).
 
 # dbplyr 2.5.0